fix: publication-script for the node-record-test executable (wip)
